Fix random road generation
The random_track script was renamed random_road to be more descriptive.

Timing lines are added to random roads each km. Previously, no timing lines were
added. This caused an assert to fail.

For circuits, make the last elevation match the first to avoid starting on a
hill.

Don't display the lap count or fastest lap if the track is not closed.