Added accelerometer to display.
Acceleration of the center of mass is returned by Rigid_Body's
acceleration method.

Move and tilt the driver's view with acceleration.

Show acceleration numbers on debug dashboard.

Don't avoid calculating slip ratio for small slip.  That caused
occasional glitches in the slopes of the cm velocity that lead to
occasional low reported acceleration.

Two_D can be constructed without width and height arguments in which
case it uses the screen size.

Position and size arguments to Two_D methods are now in percent of width or height.

Allow inverted rectangles.