Various minor fixes
Ignore generated files.

Put a timing line in the skid pad so the timing inforamtion is correct.  Added a
comment to Timing_Info describing the root of the problem.  No good solution
yet.

Record less data for replay.

Call SDL_QUIT explicitly to make sure the window closses on quit.