Fix lap time display during practice
Add Timing_Info::current_lap_time() for the elapsed time on the current, incomplete,
lap. Add a lap argument to lap_time() and return the elapsed time to the end of that lap.

Return the segment by reference instead of pointer from Road::segment_at().

On the "D" tracks, make the timing lines < track length.