Finished reading car definition file.
read_car() is now a friend of Car. Removed Car_Reader class.

Re-construct the chassis to clear the particle vectors before reading. This avoids the
need for a method to do that.

Fixed trailing 0 in on-screen gear display.

Adjusted car files slightly. Renamed peak-engine-rpm tag peak-rpm.

Removed media/xml_parser files. Added xml.cc and .h for helper functions

Fixed placement of braking markers on curves. Issue #7.

Fixed determination of track position for finding the surface being driven on. This caused
the car to act like it's in the grass when it's on the kerb.

Pass gravity to the car in World::interact(). This avoids the need to set it when reading
the car file.