Fix dim cars on some tracks.
Explicitly disable textures in models that aren't textured. This fixes a problem
with cars being rendered almost black. This happens then the last texture used
for the track was the guardrail. The dark edge of the texture image combined
with the car color. If the last track texture was a wall, it didn't look so bad
because it was a light texture combined with the car color.

If can interaction is disabled (--no-interaction), start all cars at the same
coordinates and don't add reaction time. This is useful for comparing performance.